requirements.txt
flask==0.12.2
requests==2.18.4
requests-oauthlib==0.8.0
watson-developer-cloud==1.0.2
urllib3==1.18.1
bs4
python3-linkedin
